Refactored lighting and rendering code to LightManager.
