+++ Core/ProjectileManager.py
